requirements.txt
Flask>=2.0.1
psycopg2-binary>=2.9.3
selenium>=3.141.0
beautifulsoup4>=4.10.0
tqdm>=4.62.3




